requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==22.1.0
autopep8==2.0.0
billiard==4.2.0
cached-property==1.5.2
celery==5.4.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.0.1
django-link-shortener==0.5
django-minio-storage==0.5.7
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
kavenegar==1.1.2
kombu==5.3.7
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
minio==7.2.0
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==2.0.3
Pillow==9.5.0
pip-upgrader==1.4.15
platformdirs==2.5.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.6
PyYAML==6.0
rabbitmq==0.2.0
redis==5.0.3
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.32.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
terminaltables==3.1.10
tomli==2.0.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
zeep==4.1.0
